fix(splitter): prevent default drag/touch behavior and extract useDragListeners hook
